Update logic for Put
