+++ GameStore/GameStore.api/games.http
@@ -1,7 +1,7 @@
 GET http://localhost:5108/games 
 
 ###
-GET http://localhost:5108/games/4
+GET http://localhost:5108/games/1
 
 ###
 POST http://localhost:5108/games
